Initial copy/paste of dcporter proposal
For my project, I want to stick as close to mbostock's master as
possible. Yet Ineed the relationship feature initiated by dcporter but
can't wait for mbostock to pull it. In addition I've applied the hack
to the original file vs d3v2.js so it may even help mbostock to include
it in his master more easily if it makes any sense